+++ src/components/Header.js
@@ -1,0 +1,14 @@
+import React from 'react';
+
+const Header = () => (
+    <div className="row">
+        <div className="col-8">
+            <header className="py-5 text-center">
+                <h2>Dyscalc</h2>
+            </header>
+        </div>
+        <div className="col-4" />
+    </div>
+)
+
+export default Header
