+++ src/components/Reset.js
@@ -1,0 +1,17 @@
+import React from 'react';
+
+const Reset = () => (
+    <div className="row">
+        <div className="offset-8 col-4">
+            <div className="row justify-content-md-center no-gutters">
+                <div id="reset" className="col-md-auto">
+                    <button className="btn btn-sm btn-outline-primary">
+                        RESET
+                    </button>
+                </div>
+            </div>
+        </div>
+    </div>
+)
+
+export default Reset
