+++ src/components/Verify.js
@@ -1,0 +1,13 @@
+import React from 'react';
+
+const Verify = () => (
+    <div className="row justify-content-md-center no-gutters">
+        <div id="verify" className="col-md-auto">
+            <button className="btn btn-sm btn-outline-primary">
+                VERIFY
+            </button>
+        </div>
+    </div>
+)
+
+export default Verify
